Refactoring repositories
